refactoring components out of home
